# todo-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "description": "A simple in-memory to-do list CRUD API — FlyRank Internship W2 A1.\n\nManage tasks with full Create / Read / Update / Delete support.\nAll data lives in memory — restarting the server resets it.\n\n**Stretch features included:** `?done=`, `?search=`, `?limit=`, `?offset=` query parameters; `GET /stats`; `POST /reset`.",
    "version": "1.0.0",
    "contact": {
      "name": "FlyRank Backend Track"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "tags": [
    { "name": "Meta",  "description": "API info and liveness" },
    { "name": "Tasks", "description": "CRUD operations on to-do tasks" },
    { "name": "Extras","description": "Stretch-goal endpoints" }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Meta"],
        "summary": "API descriptor",
        "description": "Returns the API name, version, and available endpoint paths.",
        "operationId": "getRoot",
        "responses": {
          "200": {
            "description": "API descriptor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiInfo" },
                "example": {
                  "name": "Task API",
                  "version": "1.0",
                  "endpoints": ["/tasks"]
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["Meta"],
        "summary": "Liveness probe",
        "description": "Returns `{\"status\": \"ok\"}`. Used by monitoring systems to confirm the server is alive.",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Health" },
                "example": { "status": "ok" }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "tags": ["Tasks"],
        "summary": "List all tasks",
        "description": "Returns the full task list. Supports optional filtering by `done` status, keyword search on `title`, and pagination via `limit` + `offset`.",
        "operationId": "getTasks",
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "description": "Filter by completion status (`true` or `false`).",
            "schema": { "type": "boolean" }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Return only tasks whose title contains this string (case-insensitive).",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of tasks to return (pagination).",
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of tasks to skip before returning results (pagination).",
            "schema": { "type": "integer", "minimum": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tasks"],
        "summary": "Create a new task",
        "description": "Creates a task with the provided `title`. The server assigns `id` (auto-incremented) and sets `done` to `false`. Returns `201 Created` with the new task.\n\nReturns `400 Bad Request` if `title` is missing or empty.",
        "operationId": "createTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTaskBody" },
              "example": { "title": "Buy milk" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" },
                "example": { "id": 4, "title": "Buy milk", "done": false }
              }
            }
          },
          "400": {
            "description": "Missing or empty title",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "title is required and must be a non-empty string" }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get a single task",
        "description": "Returns the task with the given `id`. Returns `404 Not Found` if no task has that id.",
        "operationId": "getTaskById",
        "parameters": [
          { "$ref": "#/components/parameters/taskId" }
        ],
        "responses": {
          "200": {
            "description": "The task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Tasks"],
        "summary": "Update a task",
        "description": "Replaces `title` and/or `done` on an existing task. At least one field is required. Returns the updated task.\n\n- `404` if the task doesn't exist.\n- `400` if the body is empty or invalid.",
        "operationId": "updateTask",
        "parameters": [
          { "$ref": "#/components/parameters/taskId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskBody" },
              "example": { "title": "Buy oat milk", "done": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": {
            "description": "Invalid or empty body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Tasks"],
        "summary": "Delete a task",
        "description": "Removes the task with the given `id`. Returns `204 No Content` on success (empty body). Returns `404` if the task doesn't exist.",
        "operationId": "deleteTask",
        "parameters": [
          { "$ref": "#/components/parameters/taskId" }
        ],
        "responses": {
          "204": {
            "description": "Task deleted — no body returned"
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "tags": ["Extras"],
        "summary": "Task statistics",
        "description": "Returns a computed summary: total tasks, how many are done, and how many are still open.",
        "operationId": "getStats",
        "responses": {
          "200": {
            "description": "Stats object",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Stats" },
                "example": { "total": 3, "done": 1, "open": 2 }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "tags": ["Extras"],
        "summary": "Reset tasks to seed data",
        "description": "Restores the original 3 example tasks and resets the id counter. Useful for demos. Demonstrates what restarting the server would also achieve — all unsaved data is lost.",
        "operationId": "resetTasks",
        "responses": {
          "200": {
            "description": "Reset confirmation with restored task list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "tasks": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Task" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "taskId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Numeric ID of the task",
        "schema": { "type": "integer", "example": 1 }
      }
    },
    "schemas": {
      "Task": {
        "type": "object",
        "required": ["id", "title", "done"],
        "properties": {
          "id":    { "type": "integer", "example": 1 },
          "title": { "type": "string",  "example": "Learn Express" },
          "done":  { "type": "boolean", "example": false }
        }
      },
      "CreateTaskBody": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "example": "Buy milk" }
        }
      },
      "UpdateTaskBody": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "Buy oat milk" },
          "done":  { "type": "boolean", "example": true }
        }
      },
      "ApiInfo": {
        "type": "object",
        "properties": {
          "name":      { "type": "string" },
          "version":   { "type": "string" },
          "endpoints": { "type": "array", "items": { "type": "string" } }
        }
      },
      "Health": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "ok" }
        }
      },
      "Stats": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "done":  { "type": "integer" },
          "open":  { "type": "integer" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Task 99 not found" }
        }
      }
    }
  }
}